refactor: dry tests
